Adicionando regras a aplicação

+++ src/main/java/com/mobile/login_auth_api/controlllers/DespesaController.java
@@ -5,6 +5,7 @@ import com.mobile.login_auth_api.domain.service.DespesaService;
 import com.mobile.login_auth_api.domain.service.UserService;
 import com.mobile.login_auth_api.domain.service.ValidateDateService;
 import com.mobile.login_auth_api.domain.user.User;
+import com.mobile.login_auth_api.dto.ApiResponse;
 import com.mobile.login_auth_api.dto.DespesaRequestDTO;
 import com.mobile.login_auth_api.dto.DespesaResponseDTO;
 import com.mobile.login_auth_api.dto.DespesaUpdateRequestDTO;
@@ -17,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.YearMonth;
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -36,14 +36,14 @@ public class DespesaController {
         try {
             User user = userService.findUser(body.userId());
             if (validateDateService.validateYearMonth(body.mesReferencia())){
-                throw new RuntimeException("Mês informado é anterior ao mês atual!");
+                throw new RuntimeException("Não é possivel criar despesa de meses anteriores ao atua!");
             }
             Despesa despesa = new Despesa(user, body.descricao(), body.valor(), body.mesReferencia());
             despesa = despesaRepository.save(despesa);
-            DespesaResponseDTO responseDTO = new DespesaResponseDTO(despesa.getId(), despesa.getDescricao(), despesa.getValor(), despesa.getMesReferencia());
+            DespesaResponseDTO responseDTO = DespesaResponseDTO.response(despesa);
             return ResponseEntity.ok(responseDTO);
         } catch (RuntimeException exception) {
-            return ResponseEntity.badRequest().body(exception.getMessage());
+            return ResponseEntity.badRequest().body(new ApiResponse(exception.getMessage()));
         }
     }
 
@@ -60,34 +60,38 @@ public class DespesaController {
             DespesaResponseDTO responseDTO = new DespesaResponseDTO(despesa.getId(), despesa.getDescricao(), despesa.getValor(), despesa.getMesReferencia());
             return ResponseEntity.ok(responseDTO);
         } catch (RuntimeException exception) {
-            return ResponseEntity.badRequest().body(exception.getMessage());
+            return ResponseEntity.badRequest().body(new ApiResponse(exception.getMessage()));
         }
     }
 
     @DeleteMapping("/remove")
     public ResponseEntity<?> removerDespesa(@RequestParam("id") String id) {
-        int update = despesaRepository.deleteDespesa(LocalDateTime.now(), id);
-        System.out.println("atualizou " + update);
-        if (update == 1) {
-            return ResponseEntity.ok("Despesa removida com sucesso!");
+        try {
+            Despesa despesa = despesaRepository.findById(id).orElseThrow(() -> new RuntimeException("Despesa não encontrada"));
+            if (validateDateService.validateYearMonth(despesa.getMesReferencia())) {
+                throw new RuntimeException("Não é possivel excluir despesas de meses anteriores ao atual!");
+            }
+            int update = despesaRepository.deleteDespesa(LocalDateTime.now(), id);
+            System.out.println("atualizou " + update);
+            if (update == 1) {
+                return ResponseEntity.ok(new ApiResponse("Despesa removida com sucesso!"));
+            }
+        } catch (RuntimeException exception) {
+            return ResponseEntity.badRequest().body(new ApiResponse(exception.getMessage()));
         }
-        return ResponseEntity.badRequest().build();
+        return ResponseEntity.internalServerError().body(new ApiResponse("Erro ao remover despesa!"));
     }
 
     @GetMapping("/get")
     public ResponseEntity<?> getDespesas(@RequestParam("userId") Long userId) {
         User user = userService.findUser(userId);
-        List<DespesaResponseDTO> lista = new ArrayList<>();
-        List<Despesa> dados = despesaRepository.findByUserAndDataExclusaoIsNull(user);
-        for (Despesa despesa : dados) {
-            lista.add(new DespesaResponseDTO(despesa.getId(), despesa.getDescricao(), despesa.getValor(), despesa.getMesReferencia()));
-        }
-        return ResponseEntity.ok().body(lista);
+        List<DespesaResponseDTO> dados = despesaRepository.findByUserAndDataExclusaoIsNull(user).stream().map(DespesaResponseDTO::response).toList();
+        return ResponseEntity.ok().body(dados);
     }
 
     @GetMapping("/{email}/{mesAno}")
-    public ResponseEntity<?> getDespesas(@PathVariable String email, @PathVariable YearMonth mesAno) {
-        return ResponseEntity.ok().body(despesaRepository.findByEmailAndMesAndNotNullDataExclusao(email, mesAno));
+    public ResponseEntity<?> getDespesas(@PathVariable String id, @PathVariable YearMonth mesAno) {
+        return ResponseEntity.ok().body(despesaRepository.findByIdAndMesAndNotNullDataExclusao(id, mesAno));
     }
 
     @GetMapping("/current-month")
